txn: Move Command's read or write process to their own file (#8296)
