+++ src/storage/mod.rs
@@ -3469,7 +3469,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, false
+                &ranges, false,
             ),
             true
         );
@@ -3481,7 +3481,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, false
+                &ranges, false,
             ),
             true
         );
@@ -3493,7 +3493,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, false
+                &ranges, false,
             ),
             false
         );
@@ -3506,7 +3506,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, false
+                &ranges, false,
             ),
             false
         );
@@ -3518,7 +3518,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, true
+                &ranges, true,
             ),
             true
         );
@@ -3530,7 +3530,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, true
+                &ranges, true,
             ),
             true
         );
@@ -3542,7 +3542,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, true
+                &ranges, true,
             ),
             false
         );
@@ -3554,7 +3554,7 @@ mod tests {
         ]);
         assert_eq!(
             <Storage<RocksEngine, DummyLockManager, DummyPdClient>>::check_key_ranges(
-                &ranges, true
+                &ranges, true,
             ),
             false
         );
