+++ src/storage/txn/commands/cleanup.rs
@@ -1,7 +1,16 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use pd_client::PdClient;
 use txn_types::{Key, TimeStamp};
+
+use crate::storage::kv::WriteData;
+use crate::storage::lock_manager::LockManager;
+use crate::storage::mvcc::MvccTxn;
+use crate::storage::txn::commands::{
+    Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteContext, WriteResult,
+};
+use crate::storage::txn::Result;
+use crate::storage::{ProcessResult, Snapshot};
 
 command! {
     /// Rollback mutations on a single key.
@@ -26,4 +35,31 @@ impl CommandExt for Cleanup {
     ts!(start_ts);
     write_bytes!(key);
     gen_lock!(key);
+}
+
+impl<S: Snapshot, L: LockManager, P: PdClient + 'static> WriteCommand<S, L, P> for Cleanup {
+    fn process_write(self, snapshot: S, context: WriteContext<'_, L, P>) -> Result<WriteResult> {
+        let mut txn = MvccTxn::new(
+            snapshot,
+            self.start_ts,
+            !self.ctx.get_not_fill_cache(),
+            context.pd_client,
+        );
+
+        let mut released_locks = ReleasedLocks::new(self.start_ts, TimeStamp::zero());
+        // The rollback must be protected, see more on
+        // [issue #7364](https://github.com/tikv/tikv/issues/7364)
+        released_locks.push(txn.cleanup(self.key, self.current_ts, true)?);
+        released_locks.wake_up(context.lock_mgr);
+
+        context.statistics.add(&txn.take_statistics());
+        let write_data = WriteData::from_modifies(txn.into_modifies());
+        Ok(WriteResult {
+            ctx: self.ctx,
+            to_be_write: write_data,
+            rows: 1,
+            pr: ProcessResult::Res,
+            lock_info: None,
+        })
+    }
 }
