+++ src/storage/txn/commands/mvcc_by_key.rs
@@ -1,8 +1,13 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::mvcc::MvccReader;
+use crate::storage::txn::commands::{
+    find_mvcc_infos_by_key, Command, CommandExt, ReadCommand, TypedCommand,
+};
+use crate::storage::txn::{ProcessResult, Result};
 use crate::storage::types::MvccInfo;
-use txn_types::Key;
+use crate::storage::{ScanMode, Snapshot, Statistics};
+use txn_types::{Key, TimeStamp};
 
 command! {
     /// Retrieve MVCC information for the given key.
@@ -24,4 +29,25 @@ impl CommandExt for MvccByKey {
     }
 
     gen_lock!(empty);
+}
+
+impl<S: Snapshot> ReadCommand<S> for MvccByKey {
+    fn process_read(self, snapshot: S, statistics: &mut Statistics) -> Result<ProcessResult> {
+        let mut reader = MvccReader::new(
+            snapshot,
+            Some(ScanMode::Forward),
+            !self.ctx.get_not_fill_cache(),
+            self.ctx.get_isolation_level(),
+        );
+        let result = find_mvcc_infos_by_key(&mut reader, &self.key, TimeStamp::max());
+        statistics.add(reader.get_statistics());
+        let (lock, writes, values) = result?;
+        Ok(ProcessResult::MvccKey {
+            mvcc: MvccInfo {
+                lock,
+                writes,
+                values,
+            },
+        })
+    }
 }
