+++ src/storage/txn/commands/mvcc_by_start_ts.rs
@@ -1,7 +1,12 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::mvcc::MvccReader;
+use crate::storage::txn::commands::{
+    find_mvcc_infos_by_key, Command, CommandExt, ReadCommand, TypedCommand,
+};
+use crate::storage::txn::{ProcessResult, Result};
 use crate::storage::types::MvccInfo;
+use crate::storage::{ScanMode, Snapshot, Statistics};
 use txn_types::{Key, TimeStamp};
 
 command! {
@@ -25,4 +30,33 @@ impl CommandExt for MvccByStartTs {
     }
 
     gen_lock!(empty);
+}
+
+impl<S: Snapshot> ReadCommand<S> for MvccByStartTs {
+    fn process_read(self, snapshot: S, statistics: &mut Statistics) -> Result<ProcessResult> {
+        let mut reader = MvccReader::new(
+            snapshot,
+            Some(ScanMode::Forward),
+            !self.ctx.get_not_fill_cache(),
+            self.ctx.get_isolation_level(),
+        );
+        match reader.seek_ts(self.start_ts)? {
+            Some(key) => {
+                let result = find_mvcc_infos_by_key(&mut reader, &key, TimeStamp::max());
+                statistics.add(reader.get_statistics());
+                let (lock, writes, values) = result?;
+                Ok(ProcessResult::MvccStartTs {
+                    mvcc: Some((
+                        key,
+                        MvccInfo {
+                            lock,
+                            writes,
+                            values,
+                        },
+                    )),
+                })
+            }
+            None => Ok(ProcessResult::MvccStartTs { mvcc: None }),
+        }
+    }
 }
