+++ src/storage/txn/commands/pause.rs
@@ -1,6 +1,15 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::kv::WriteData;
+use crate::storage::lock_manager::LockManager;
+use crate::storage::txn::commands::{
+    Command, CommandExt, TypedCommand, WriteCommand, WriteContext, WriteResult,
+};
+use crate::storage::txn::Result;
+use crate::storage::{ProcessResult, Snapshot};
+use pd_client::PdClient;
+use std::thread;
+use std::time::Duration;
 use txn_types::Key;
 
 command! {
@@ -23,4 +32,17 @@ impl CommandExt for Pause {
     tag!(pause);
     write_bytes!(keys: multiple);
     gen_lock!(keys: multiple);
+}
+
+impl<S: Snapshot, L: LockManager, P: PdClient + 'static> WriteCommand<S, L, P> for Pause {
+    fn process_write(self, _snapshot: S, _context: WriteContext<'_, L, P>) -> Result<WriteResult> {
+        thread::sleep(Duration::from_millis(self.duration));
+        Ok(WriteResult {
+            ctx: self.ctx,
+            to_be_write: WriteData::default(),
+            rows: 0,
+            pr: ProcessResult::Res,
+            lock_info: None,
+        })
+    }
 }
