+++ src/storage/txn/commands/prewrite.rs
@@ -1,11 +1,23 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
+use engine_traits::CF_WRITE;
+use pd_client::PdClient;
+use txn_types::{Key, Mutation, TimeStamp};
+
+use crate::storage::kv::WriteData;
+use crate::storage::lock_manager::LockManager;
+use crate::storage::mvcc::{
+    has_data_in_range, Error as MvccError, ErrorInner as MvccErrorInner, MvccTxn,
+};
+use crate::storage::txn::commands::{WriteCommand, WriteContext, WriteResult};
+use crate::storage::txn::{Error, Result};
 use crate::storage::{
     txn::commands::{Command, CommandExt, TypedCommand},
     types::PrewriteResult,
-    Context,
+    Context, Error as StorageError, ProcessResult, ScanMode, Snapshot,
 };
-use txn_types::{Mutation, TimeStamp};
+
+pub(crate) const FORWARD_MIN_MUTATIONS_NUM: usize = 12;
 
 command! {
     /// The prewrite phase of a transaction. The first phase of 2PC.
@@ -115,5 +127,322 @@ impl Prewrite {
             None,
             ctx,
         )
+    }
+}
+
+impl<S: Snapshot, L: LockManager, P: PdClient + 'static> WriteCommand<S, L, P> for Prewrite {
+    fn process_write(
+        mut self,
+        snapshot: S,
+        context: WriteContext<'_, L, P>,
+    ) -> Result<WriteResult> {
+        let mut scan_mode = None;
+        let rows = self.mutations.len();
+        if rows > FORWARD_MIN_MUTATIONS_NUM {
+            self.mutations.sort_by(|a, b| a.key().cmp(b.key()));
+            let left_key = self.mutations.first().unwrap().key();
+            let right_key = self
+                .mutations
+                .last()
+                .unwrap()
+                .key()
+                .clone()
+                .append_ts(TimeStamp::zero());
+            if !has_data_in_range(
+                snapshot.clone(),
+                CF_WRITE,
+                left_key,
+                &right_key,
+                &mut context.statistics.write,
+            )? {
+                // If there is no data in range, we could skip constraint check, and use Forward seek for CF_LOCK.
+                // Because in most instances, there won't be more than one transaction write the same key. Seek
+                // operation could skip nonexistent key in CF_LOCK.
+                self.skip_constraint_check = true;
+                scan_mode = Some(ScanMode::Forward)
+            }
+        }
+        let mut txn = if scan_mode.is_some() {
+            MvccTxn::for_scan(
+                snapshot,
+                scan_mode,
+                self.start_ts,
+                !self.ctx.get_not_fill_cache(),
+                context.pd_client,
+            )
+        } else {
+            MvccTxn::new(
+                snapshot,
+                self.start_ts,
+                !self.ctx.get_not_fill_cache(),
+                context.pd_client,
+            )
+        };
+
+        // Set extra op here for getting the write record when check write conflict in prewrite.
+        txn.extra_op = context.extra_op;
+
+        let primary_key = Key::from_raw(&self.primary);
+        let mut locks = vec![];
+        let mut async_commit_ts = TimeStamp::zero();
+        for m in self.mutations {
+            let mut secondaries = &self.secondary_keys.as_ref().map(|_| vec![]);
+
+            if m.key() == &primary_key {
+                secondaries = &self.secondary_keys;
+            }
+            match txn.prewrite(
+                m,
+                &self.primary,
+                secondaries,
+                self.skip_constraint_check,
+                self.lock_ttl,
+                self.txn_size,
+                self.min_commit_ts,
+            ) {
+                Ok(ts) => {
+                    if secondaries.is_some() {
+                        async_commit_ts = ts;
+                    }
+                }
+                e @ Err(MvccError(box MvccErrorInner::KeyIsLocked { .. })) => {
+                    locks.push(
+                        e.map(|_| ())
+                            .map_err(Error::from)
+                            .map_err(StorageError::from),
+                    );
+                }
+                Err(e) => return Err(Error::from(e)),
+            }
+        }
+
+        context.statistics.add(&txn.take_statistics());
+        let (pr, to_be_write, rows, ctx, lock_info) = if locks.is_empty() {
+            let pr = ProcessResult::PrewriteResult {
+                result: PrewriteResult {
+                    locks: vec![],
+                    min_commit_ts: async_commit_ts,
+                },
+            };
+            let txn_extra = txn.take_extra();
+            let write_data = WriteData::new(txn.into_modifies(), txn_extra);
+            (pr, write_data, rows, self.ctx, None)
+        } else {
+            // Skip write stage if some keys are locked.
+            let pr = ProcessResult::PrewriteResult {
+                result: PrewriteResult {
+                    locks,
+                    min_commit_ts: async_commit_ts,
+                },
+            };
+            (pr, WriteData::default(), 0, self.ctx, None)
+        };
+        Ok(WriteResult {
+            ctx,
+            to_be_write,
+            rows,
+            pr,
+            lock_info,
+        })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use kvproto::kvrpcpb::{Context, ExtraOp};
+
+    use engine_traits::CF_WRITE;
+    use pd_client::DummyPdClient;
+    use txn_types::TimeStamp;
+    use txn_types::{Key, Mutation};
+
+    use crate::storage::mvcc::{Error as MvccError, ErrorInner as MvccErrorInner};
+    use crate::storage::txn::commands::{
+        Commit, Prewrite, WriteContext, FORWARD_MIN_MUTATIONS_NUM,
+    };
+    use crate::storage::txn::LockInfo;
+    use crate::storage::txn::{Error, ErrorInner, Result};
+    use crate::storage::DummyLockManager;
+    use crate::storage::{
+        Engine, PrewriteResult, ProcessResult, Snapshot, Statistics, TestEngineBuilder,
+    };
+
+    fn inner_test_prewrite_skip_constraint_check(pri_key_number: u8, write_num: usize) {
+        let mut mutations = Vec::default();
+        let pri_key = &[pri_key_number];
+        for i in 0..write_num {
+            mutations.push(Mutation::Insert((
+                Key::from_raw(&[i as u8]),
+                b"100".to_vec(),
+            )));
+        }
+        let mut statistic = Statistics::default();
+        let engine = TestEngineBuilder::new().build().unwrap();
+        prewrite(
+            &engine,
+            &mut statistic,
+            vec![Mutation::Put((
+                Key::from_raw(&[pri_key_number]),
+                b"100".to_vec(),
+            ))],
+            pri_key.to_vec(),
+            99,
+        )
+        .unwrap();
+        assert_eq!(1, statistic.write.seek);
+        let e = prewrite(
+            &engine,
+            &mut statistic,
+            mutations.clone(),
+            pri_key.to_vec(),
+            100,
+        )
+        .err()
+        .unwrap();
+        assert_eq!(2, statistic.write.seek);
+        match e {
+            Error(box ErrorInner::Mvcc(MvccError(box MvccErrorInner::KeyIsLocked(_)))) => (),
+            _ => panic!("error type not match"),
+        }
+        commit(
+            &engine,
+            &mut statistic,
+            vec![Key::from_raw(&[pri_key_number])],
+            99,
+            102,
+        )
+        .unwrap();
+        assert_eq!(2, statistic.write.seek);
+        let e = prewrite(
+            &engine,
+            &mut statistic,
+            mutations.clone(),
+            pri_key.to_vec(),
+            101,
+        )
+        .err()
+        .unwrap();
+        match e {
+            Error(box ErrorInner::Mvcc(MvccError(box MvccErrorInner::WriteConflict {
+                ..
+            }))) => (),
+            _ => panic!("error type not match"),
+        }
+        let e = prewrite(
+            &engine,
+            &mut statistic,
+            mutations.clone(),
+            pri_key.to_vec(),
+            104,
+        )
+        .err()
+        .unwrap();
+        match e {
+            Error(box ErrorInner::Mvcc(MvccError(box MvccErrorInner::AlreadyExist { .. }))) => (),
+            _ => panic!("error type not match"),
+        }
+
+        statistic.write.seek = 0;
+        let ctx = Context::default();
+        engine
+            .delete_cf(
+                &ctx,
+                CF_WRITE,
+                Key::from_raw(&[pri_key_number]).append_ts(102.into()),
+            )
+            .unwrap();
+        prewrite(
+            &engine,
+            &mut statistic,
+            mutations.clone(),
+            pri_key.to_vec(),
+            104,
+        )
+        .unwrap();
+        // All keys are prewrited successful with only one seek operations.
+        assert_eq!(1, statistic.write.seek);
+        let keys: Vec<Key> = mutations.iter().map(|m| m.key().clone()).collect();
+        commit(&engine, &mut statistic, keys.clone(), 104, 105).unwrap();
+        let snap = engine.snapshot(&ctx).unwrap();
+        for k in keys {
+            let v = snap.get_cf(CF_WRITE, &k.append_ts(105.into())).unwrap();
+            assert!(v.is_some());
+        }
+    }
+
+    #[test]
+    fn test_prewrite_skip_constraint_check() {
+        inner_test_prewrite_skip_constraint_check(0, FORWARD_MIN_MUTATIONS_NUM + 1);
+        inner_test_prewrite_skip_constraint_check(5, FORWARD_MIN_MUTATIONS_NUM + 1);
+        inner_test_prewrite_skip_constraint_check(
+            FORWARD_MIN_MUTATIONS_NUM as u8,
+            FORWARD_MIN_MUTATIONS_NUM + 1,
+        );
+    }
+
+    fn prewrite<E: Engine>(
+        engine: &E,
+        statistics: &mut Statistics,
+        mutations: Vec<Mutation>,
+        primary: Vec<u8>,
+        start_ts: u64,
+    ) -> Result<()> {
+        let ctx = Context::default();
+        let snap = engine.snapshot(&ctx)?;
+        let cmd = Prewrite::with_defaults(mutations, primary, TimeStamp::from(start_ts));
+        let context = WriteContext {
+            lock_mgr: &DummyLockManager {},
+            pd_client: Arc::new(DummyPdClient::new()),
+            extra_op: ExtraOp::Noop,
+            statistics,
+            pipelined_pessimistic_lock: false,
+        };
+        let ret = cmd.cmd.process_write(snap, context)?;
+        if let ProcessResult::PrewriteResult {
+            result: PrewriteResult { locks, .. },
+        } = ret.pr
+        {
+            if !locks.is_empty() {
+                let info = LockInfo::default();
+                return Err(Error::from(ErrorInner::Mvcc(MvccError::from(
+                    MvccErrorInner::KeyIsLocked(info),
+                ))));
+            }
+        }
+        let ctx = Context::default();
+        engine.write(&ctx, ret.to_be_write).unwrap();
+        Ok(())
+    }
+
+    fn commit<E: Engine>(
+        engine: &E,
+        statistics: &mut Statistics,
+        keys: Vec<Key>,
+        lock_ts: u64,
+        commit_ts: u64,
+    ) -> Result<()> {
+        let ctx = Context::default();
+        let snap = engine.snapshot(&ctx)?;
+        let cmd = Commit::new(
+            keys,
+            TimeStamp::from(lock_ts),
+            TimeStamp::from(commit_ts),
+            ctx,
+        );
+
+        let context = WriteContext {
+            lock_mgr: &DummyLockManager {},
+            pd_client: Arc::new(DummyPdClient::new()),
+            extra_op: ExtraOp::Noop,
+            statistics,
+            pipelined_pessimistic_lock: false,
+        };
+
+        let ret = cmd.cmd.process_write(snap, context)?;
+        let ctx = Context::default();
+        engine.write(&ctx, ret.to_be_write).unwrap();
+        Ok(())
     }
 }
