+++ src/storage/txn/commands/resolve_lock_lite.rs
@@ -1,6 +1,14 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::kv::WriteData;
+use crate::storage::lock_manager::LockManager;
+use crate::storage::mvcc::MvccTxn;
+use crate::storage::txn::commands::{
+    Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteContext, WriteResult,
+};
+use crate::storage::txn::Result;
+use crate::storage::{ProcessResult, Snapshot};
+use pd_client::PdClient;
 use txn_types::{Key, TimeStamp};
 
 command! {
@@ -25,4 +33,38 @@ impl CommandExt for ResolveLockLite {
     command_method!(is_sys_cmd, bool, true);
     write_bytes!(resolve_keys: multiple);
     gen_lock!(resolve_keys: multiple);
+}
+
+impl<S: Snapshot, L: LockManager, P: PdClient + 'static> WriteCommand<S, L, P> for ResolveLockLite {
+    fn process_write(self, snapshot: S, context: WriteContext<'_, L, P>) -> Result<WriteResult> {
+        let mut txn = MvccTxn::new(
+            snapshot,
+            self.start_ts,
+            !self.ctx.get_not_fill_cache(),
+            context.pd_client,
+        );
+
+        let rows = self.resolve_keys.len();
+        // ti-client guarantees the size of resolve_keys will not too large, so no necessary
+        // to control the write_size as ResolveLock.
+        let mut released_locks = ReleasedLocks::new(self.start_ts, self.commit_ts);
+        for key in self.resolve_keys {
+            released_locks.push(if !self.commit_ts.is_zero() {
+                txn.commit(key, self.commit_ts)?
+            } else {
+                txn.rollback(key)?
+            });
+        }
+        released_locks.wake_up(context.lock_mgr);
+
+        context.statistics.add(&txn.take_statistics());
+        let write_data = WriteData::from_modifies(txn.into_modifies());
+        Ok(WriteResult {
+            ctx: self.ctx,
+            to_be_write: write_data,
+            rows,
+            pr: ProcessResult::Res,
+            lock_info: None,
+        })
+    }
 }
