+++ src/storage/txn/commands/rollback.rs
@@ -1,6 +1,14 @@
 // Copyright 2020 TiKV Project Authors. Licensed under Apache-2.0.
 
-use crate::storage::txn::commands::{Command, CommandExt, TypedCommand};
+use crate::storage::kv::WriteData;
+use crate::storage::lock_manager::LockManager;
+use crate::storage::mvcc::MvccTxn;
+use crate::storage::txn::commands::{
+    Command, CommandExt, ReleasedLocks, TypedCommand, WriteCommand, WriteContext, WriteResult,
+};
+use crate::storage::txn::Result;
+use crate::storage::{ProcessResult, Snapshot};
+use pd_client::PdClient;
 use txn_types::{Key, TimeStamp};
 
 command! {
@@ -23,4 +31,32 @@ impl CommandExt for Rollback {
     ts!(start_ts);
     write_bytes!(keys: multiple);
     gen_lock!(keys: multiple);
+}
+
+impl<S: Snapshot, L: LockManager, P: PdClient + 'static> WriteCommand<S, L, P> for Rollback {
+    fn process_write(self, snapshot: S, context: WriteContext<'_, L, P>) -> Result<WriteResult> {
+        let mut txn = MvccTxn::new(
+            snapshot,
+            self.start_ts,
+            !self.ctx.get_not_fill_cache(),
+            context.pd_client,
+        );
+
+        let rows = self.keys.len();
+        let mut released_locks = ReleasedLocks::new(self.start_ts, TimeStamp::zero());
+        for k in self.keys {
+            released_locks.push(txn.rollback(k)?);
+        }
+        released_locks.wake_up(context.lock_mgr);
+
+        context.statistics.add(&txn.take_statistics());
+        let write_data = WriteData::from_modifies(txn.into_modifies());
+        Ok(WriteResult {
+            ctx: self.ctx,
+            to_be_write: write_data,
+            rows,
+            pr: ProcessResult::Res,
+            lock_info: None,
+        })
+    }
 }
